+++ HISTORY.rst
@@ -2,19 +2,21 @@
 History
 =======
 
-0.3.2 (XXXX-XX-XX)
+0.3.2 (2021-XX-XX)
 ------------------
-
+* Refactorize unit tests
 * Add "naive" and "top-k" methods in MapieClassifier
 * Include J+aB method in regression tutorial
 * Add MNIST example for classification
 * Add replication of the Chen Xu's tutorial testing Jackknife+aB vs Jackknife+.
+* Add Jackknife+-after-Bootstrap documentation
 
 0.3.1 (2021-11-19)
 ------------------
 
 * Add Jackknife+-after-Bootstrap method and add mean and median as aggregation functions
 * Add "cumulative_score" method in MapieClassifier
+* Allow image as input in MapieClassifier
 
 0.3.0 (2021-09-10)
 ------------------